Fix RequestTrackerImpl uncontrolled destruction.
Objective-C blocks' lifetime can't be precisely controlled when they
are in an autoreleased pool. Since moving to ARC, base::BindBlock will
cause blocks to be added to such a pool.

This will affect in turn the objects retained by the pool. In this
case, the object is a scoped_refptr and controlling exactly its
lifetime is extremely important since some work is done in its
destructor (the RequestTrackerImpl object is removed from a global
map).

BUG=

Review-Url: https://codereview.chromium.org/2394143002
Cr-Original-Commit-Position: refs/heads/master@{#423559}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0c8e198322da1b90ee4899e0755e97e81deb6e9a