Enable lost URL detection in the Physical Web scanner
The Physical Web scanner was initially designed to run for a relatively
short duration (less than a minute) and would report any URLs seen during
this time as nearby. With this change, the scanner will keep track of how
recently a URL was seen so that URLs that are no longer nearby can be
removed from the list.

BUG=630769

Review-Url: https://chromiumcodereview.appspot.com/2413923002
Cr-Original-Commit-Position: refs/heads/master@{#426859}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7afb153e71eee978830f552c3002348b9c9501d5