Add WebStateObserver observing the progress of page loads.
This CL adds an observer observing the progress of page loads coming from the reading list.
If the page loads slowly, attemps to load an offline version of the page (if possible).

Companion downstream CL that shows how the WSO would be added:
https://chromereviews.googleplex.com/533417013

BUG=None.

Review-Url: https://codereview.chromium.org/2480233005
Cr-Original-Commit-Position: refs/heads/master@{#431858}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71e59eecddca6c2a909bd6ade0b23951530a2b54