[iOS] Upstream web_controller_provider_impl.{h,mm}.
Upstream the implementation of WebControllerProviderImpl, split the
file to move the factory and the class in separate files. Fix deps.

Dependency on the upstream class can be added in that CL as it is in
a different namespace as the downstream implementation and do not
conflict.

BUG=None

Review-Url: https://codereview.chromium.org/2527123003
Cr-Original-Commit-Position: refs/heads/master@{#434485}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 21a52db5fec1a61393b2aedbe5bf0b834d9d97eb