Adds a new constructor for gfx::Image on iOS.
Chrome iOS is moving to Objective-C Automatic Reference Counting. Under
it, arguments are supposed to be autoreleased. This CL adds a new
constructor that can take both a retained and an autoreleased object.
This is a part of a multiple CL setup that will change the Image(UIImage*) constructor from consuming +1 refcount to retaining the argument.

TEST=none
BUG=645125

Review-Url: https://codereview.chromium.org/2504683002
Cr-Original-Commit-Position: refs/heads/master@{#437513}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90031d9584452e5018e68a3ad2ff47634aab70b6