Modify StartPhysicalWebDiscovery to take an ios::ChromeBrowserState param
In a future CL, the Physical Web's auto-optin behavior will be modified to
require that Google is configured as the default search provider. This CL
adds a new overload for StartPhysicalWebDiscovery so downstream users can
be modified to use the new signature that provides access to the browser
state, from which we can determine the default search provider.

BUG=630769

Review-Url: https://codereview.chromium.org/2566753002
Cr-Original-Commit-Position: refs/heads/master@{#438003}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7d1e59e370b8810d5ba953d38887ed4a79af4a5e