Modify Physical Web auto-optin behavior on Chrome for iOS
The Physical Web feature will auto-enable if certain conditions are met
when the browser first launches or switches modes. This CL adds a new
condition: Google must be configured as the default search provider.

The behavior when the auto-enable check fails is also changed. Previously,
if the check failed the preference would be left in the default/onboarding
state. With this CL, the preference will be explicitly disabled if the
check fails.

BUG=630769

Review-Url: https://codereview.chromium.org/2573923002
Cr-Original-Commit-Position: refs/heads/master@{#438898}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40351cf5845e48c03a54f7a2873955e4c68c6c44