Initialize the BrowserContainerView with the right frame.
When the iPad Tab Switcher creates an incognito tab for the first time, an incognito
BVC is created.
The Tab Switcher takes a screenshot of the new incognito tab, and animates its size
from a small cell to a fullscreen tab.
Before this CL, the first incognito tab was created with an incorrect size, resulting in
the screenshot used in the animation being incorrect.

BUG=671724

Review-Url: https://codereview.chromium.org/2591683002
Cr-Original-Commit-Position: refs/heads/master@{#439792}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50d5ba259ac6c76e63884412fb1a214858b81654