Remove image name
This CL removes the image name and add a unique filename with a default PNG
extension. Before this CL the images were named with their filename. But in
order to be saved to the device, the file should have a valid extension.

Revert of the resolution of 656641

BUG=656641, 674555
TEST=Download images with different image type (gif, tiff, bmp, jpg, png)
and verify it is working as expected.

Review-Url: https://codereview.chromium.org/2595573002
Cr-Original-Commit-Position: refs/heads/master@{#440080}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16cfe1ff158f1190e1b26eb1e491409ae904a9b8