[ios] Fixed web-compat bug related to resetting top content inset.
OverscrollActionsController resets top content inset every time when
scolling is finished. Before this change it assumed that bottom inset
should be always reset to 0. This is not correct for cases when soft
keyboard is displayed and bottom offset is non-zero to accomodate the
keyboard.

This CL adds resetScrollViewTopContentInset method and calls it when
scrolling is finished. Remaining fixed will be landed in separate CLs
to make changes less risky.

BUG=669908,675763

Review-Url: https://codereview.chromium.org/2589973004
Cr-Original-Commit-Position: refs/heads/master@{#440273}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0774dcc15d6c7f59bcca77b5c2d1288b49911707