Reload offline version on load failure
Change the way we trigger offline version.
The ReadingListWebStateObserver now listens for all page loads.
If there is an error or the page is too slow, and if an offline page is
available, show the native content instead.
Offline page is displayed by reloading on the offline URL.

BUG=671954, 671160, 671963

Review-Url: https://codereview.chromium.org/2578973002
Cr-Original-Commit-Position: refs/heads/master@{#440367}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d91357e728e0d6821b6fed7fdce60d3dfc95193