[ios] Removed CookieStoreIOS::UnSynchronize.
This method was called when the last incognito tab was closed to prevent
Synchronizing incognito cookies with shared cookie storage. WKWebView's
incognito is ephemeral and web view cookies are not stored anywhere and
not accessible for Chrome app.

Also removed SwitchSynchronizedStore which was used only in tests and
made SetSynchronizedWithSystemStore public to keep it using in tests for now
(instead of UnSynchronize and SwitchSynchronizedStore). Once most of
CookieStoreIOS methods are removed it should be easier to remove
SetSynchronizedWithSystemStore as well.

BUG=676144

Review-Url: https://codereview.chromium.org/2597933003
Cr-Original-Commit-Position: refs/heads/master@{#440638}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 61488b67c9a2cf7fbe27b49e312531bb80af8f13