[ios] Added a CHECK to didCommitNavigation: to debug crbug.com/676340.
didCommitNavigation: crashes if NavigationManager is empty. This new
CHECK crashes if MN is empty and |didStartProvisionalNavigation| was
called.

If this CHECK is hit it means that |didStartProvisionalNavigation| was
called as expected but it did not call |registerLoadRequest|. If that's
the case then |didStartProvisionalNavigation| needs to be changed to
always call |registerLoadRequest| if MN is empty.

BUG=676458

Review-Url: https://codereview.chromium.org/2596243003
Cr-Original-Commit-Position: refs/heads/master@{#440796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 272901feb7b3ed31f1aa50e50b799512f1152917