[ios] Log UMA if didStartProvisionalNavigation doesn't add pending item.
Chrome for iOS crashes in |didCommitNavigation:| with empty Navigation
Manager. It's unclear how app can get into this state and one assumption
is that didStartProvisionalNavigation: does not add a pending navigation
item. This UMA histogram is logged when didStartProvisionalNavigation:
fails to add a pending item.

This histogram is added only to verify the assumption and should be removed
after M46 is released.

BUG=676458

Review-Url: https://codereview.chromium.org/2602133002
Cr-Original-Commit-Position: refs/heads/master@{#442047}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6fef8190db76bbc5e9716d46296bfda636fad7f