[ios] Extended webStateDidLoadPage with withSuccess flag.
webStateDidLoadPage is supposed to be an Objective-C adapter for
WebStateObserver::PageLoaded(web::PageLoadCompletionStatus), so this
BOOL flag is needed to mirror C++ method.

Also webStateDidLoadPage will be a replacement for
-[CRWWebDelegate webDidFinishWithURL:loadSuccess:] which already has
this BOOL flag.

C++ version of this menthod uses web::PageLoadCompletionStatus enum,
which is not needed for webStateDidLoadPage: because Objective-C is
very verbose language and withSuccess: suffix perfectly explains the
purpose of the argument.

BUG=675001

Review-Url: https://codereview.chromium.org/2619103003
Cr-Original-Commit-Position: refs/heads/master@{#442283}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15eecbc612e29f171233318b7b0ff227b8e0ed5f