Mojo JS bindings: make sure that console.log() could be used in the bindings.
It is convenient for debugging.

Although console.log() is supported by chrome directly, mojo bindings currently are loaded using gin. Therefore, console.log() is not always available and needs to be defined as a module. After we move away from gin/AMD, we should be able to get rid of this.

BUG=579646
TBR=jochen@chromium.org
(deferring to eugenebut)

Review-Url: https://codereview.chromium.org/2618693004
Cr-Original-Commit-Position: refs/heads/master@{#442972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ff21f95261daa80eeff96b7df798da9b85b5eee