Add more criteria to wait until keyboard appears.
The testBrowsingPostEntryWithKeyboard test is flaky on device, with the
failure that the Go button on the keyboard is not interactable when tap
is performed. This implies that the tap is reached before the keyboard
is fully interactable, so this CL adds a requirement that the button is
interactable to the wait criteria.

BUG=670380

Review-Url: https://codereview.chromium.org/2626143002
Cr-Original-Commit-Position: refs/heads/master@{#443038}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0ba8c7a2c6f419d37532d31c79d2eaf62723ce8