[Signin iOS] Ensure SigninInteractionController displays signin errors.
SigninInteractionController uses an AlertCoordinator to display
potential sign-in errors, which requires the view controler to use to
present the alert dialog.
To ensure SigninInteractionController can always display correctly
errors, the AlertCoordinator must be given the top most visible view
controller (some errors are displayed before the sign-in screen is
shown, and some are displayed on top of the sign-in screen).

BUG=679809

Review-Url: https://codereview.chromium.org/2626003004
Cr-Original-Commit-Position: refs/heads/master@{#443230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f9dd03c83e9a69b48a941b553fe255c68d401d6