Do not start timer on when reloading a Reading List offline page on iOS.
If a user has a distilled entry in the Reading List and is on a poor network, the
distilled entry will be loaded and the user cannot access the online page.
To avoid blocking the user from the up to date page, reloading a page from the
offline page will not start the offline triggering timer.
Reloading again from the pending loading will load back the online version.

BUG=680481

Review-Url: https://codereview.chromium.org/2616093002
Cr-Original-Commit-Position: refs/heads/master@{#443287}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3dcbdfb6ef790697a538ef3f0ab909e44af8ee5d