Allow using a BrowserViewController without BookmarkModel for testing.
Loading the bookmarks introduce some synchronisation point (as it is
done in a background sequence) which may be the causing the flakes
of BrowserViewWranglerTest.TestInitNilObserver.

Change BrowserViewController to works if the BookmarkModel has not
been created and change the test to not create it.

BUG=666816

Review-Url: https://codereview.chromium.org/2632393002
Cr-Original-Commit-Position: refs/heads/master@{#444047}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc432f4211bfe92a50e6766d62ed8f833ab0c42d