Adjust disabling of synchronization and add wait to test.
The testBrowsingPostEntryWithKeyboard test is flaky on iPads on iOS 9.
This CL addresses the problem by ensuring that synchronization is
disabled and that the web view is interactable before the tap on the
text field occurs. It also reenables the test, since failures appear only
on main tree.

BUG=670380

Review-Url: https://codereview.chromium.org/2642523002
Cr-Original-Commit-Position: refs/heads/master@{#444925}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1d36aa302699a66f20326ae8589cde2ca9a103e