Store the distilled_url in Reading List entry in Reading List on iOS.
The URL of the distilled page is not necessarily the URL of the Reading
List entry (in case of redirection).
Store the URL of the page that was distilled in the database.
This CL adds the field in ReadingListEntry.
This field will be used in a next CL.

BUG=682666

Review-Url: https://codereview.chromium.org/2647763005
Cr-Original-Commit-Position: refs/heads/master@{#445063}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 695d0c65669b7af00f8daec96d497519c0e0cb46