Store distilled URL during distillation in Reading List on iOS
The URL of the distilled page is not necessarily the URL of the Reading
List entry (incase of redirection).
If the distillation is done on a redirected page, store it along the
distilled_path.

I also filed crbug.com/682705 to have the "redirected_url" directly in the proto file
and avoid having to take it via callback.

BUG=682666

Review-Url: https://codereview.chromium.org/2646993002
Cr-Original-Commit-Position: refs/heads/master@{#445071}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69e7e3b844f1cedf53701f07bf0d3219fc2d3176