[Sync] Make SMTP always have the dump stack function.
Previously this was threaded through ModelTypeController, which is bad
because it meant we uploaded completely useless stacks for any errors
in the model code that occurred prior to sync starting up.

Now the function is curried directly into the processor factory
function, so we always have access to it.

BUG=673883,680093

Review-Url: https://codereview.chromium.org/2639393002
Cr-Original-Commit-Position: refs/heads/master@{#445122}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f90a97e73ca55f5bbe2fd23fc67da6c3c92010f