[ios] Fix Tools Menu presentation
After the Tab Grid commit, the Menu wasn't being displayed anymore.
In order to fix it I made TabStripContainerVC conform to MenuPresentationDelegate
and ZoomTransitionDelegate.

The menu's default presentation (at the middle of the screen) wasn't working.
So this could be just a workaround.

BUG=none

Review-Url: https://codereview.chromium.org/2639643003
Cr-Original-Commit-Position: refs/heads/master@{#445248}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eea06f79f570e58dfe52f788027cc6d780dbc470