WebStateList is an array of web::WebState* wrappers.
Introduce a super class of TabModel with the goal of moving all
the code dealing with Tab management (iteration, selection, ...)
eventually.

This is a first step in removing the Tab* class that is mostly
used as an Objective-C wrapper of web::WebState* with lots of
helper methods.

BUG=681867

Review-Url: https://codereview.chromium.org/2640093004
Cr-Original-Commit-Position: refs/heads/master@{#445362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3579f53085d00ff7a80eb97bb644a90becddc068