Add the suggestions UI in chrome
This CL adds an entry to the toolbar menu to enter the suggestions UI.
The position in the toolbar menu is not ideal but it is useful to test the UI
in the context of Chrome. It will be changed.
This entry is accessible using an experimental setting.

BUG=679369

Review-Url: https://codereview.chromium.org/2643753002
Cr-Original-Commit-Position: refs/heads/master@{#445984}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2e44fbf6b351a10a698b1571019fe0a06bf132c