Reading List IOS: save PDF for offline display.
The ReadingListDistillerPage now reports the mime type of the downloaded page.
If the mime type is not text/html, the distillation will fail, but the caller (URLDownloader)
has the opportunity to save the file.

URLDownloader uses URLFetcher to save the files on disk and show them while
offline.

BUG=678112

Review-Url: https://codereview.chromium.org/2647883003
Cr-Original-Commit-Position: refs/heads/master@{#445988}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c34727b3b457aa0cd30e51928a532677a700c6f4