Discard pending navigation item when PassKit file is downloaded.
Discard the pending navigation item when the browser navigates to a
PassKit file, so that the item is not persisted beyond the completion
of the download. Do not discard when the previous entry is loaded in
a native view, to avoid native view to web view to native view bugs.

BUG=672521

Review-Url: https://codereview.chromium.org/2570173002
Cr-Original-Commit-Position: refs/heads/master@{#446167}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0986378bcdb6e01f0cdb1dc5a828a7873cda144