Update Physical Web omnibox icons to match Chrome's theme
An icon with the Physical Web logo is shown on the right end of Physical
Web omnibox suggestions. Previously, these icons used colors that looked
out of place with the other omnibox icons. This CL replaces the PW icons
with new ones that match the theme.

BUG=685343

Review-Url: https://codereview.chromium.org/2645403010
Cr-Original-Commit-Position: refs/heads/master@{#446206}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9896346c3c7b07539e00668b9967a6e714dfeeac