Prevent crash at shutdown if no RequestTracker instance was created.
If no RequestTracker was created (which could happen if the user
never follow the sign-in workflow based on UIWebView) the globals
are not initialized which cause a crash.

Invokes the initialisation of the globals in the cleanup method
(it cannot be remove yet as RequestTracker is still used by the
sign-in workflow based on UIWebView).

BUG=684611

Review-Url: https://codereview.chromium.org/2664243002
Cr-Original-Commit-Position: refs/heads/master@{#447509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8a738dfc5b32886bab40ad098aaa087de29f3b06