Add ContentSuggestionsMediator
This CL adds a Mediator level in ContentSuggestions.
The mediator classes are used to create a interface between the data source in
C++ and the UI in Objective-C.
This CL only creates the classes to see the interactions from an architectural point.
Most of the methods are no-op for now.

BUG=679369

Review-Url: https://codereview.chromium.org/2638413006
Cr-Original-Commit-Position: refs/heads/master@{#447511}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ac7f3e7f45b25cf8b3283a7df27bf68b3a72157