[ios] Automatically fail any WebTest whose render process crashes.
We speculate that some tests are flaky because iOS is killing WKWebViews while
the tests are running.  This CL adds a GlobalWebStateObserver method to listen
for this and explicitly fail the current test when a WKWebView dies.  Having an
explicit failure will make these cases more apparent when looking through test
logs.

This CL also adds an escape hatch for tests that intentionally crash the
renderer.

BUG=None

Review-Url: https://codereview.chromium.org/2666953003
Cr-Original-Commit-Position: refs/heads/master@{#447615}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: caeca2dde86a85efeda94866988b3659f5951889