[ios clean] Remove stop on Browser coordinator dealloc
The previous code caused that dismissing the ToolMenu would dismiss the whole Tab
(Toolbar+content view) and left the App in a corrupted state. (The coordinators were
not stopped but their VC was dismissed).

This is a temporary fix. Marq will submit a more robust solution to this problem later on.

BUG=

Review-Url: https://codereview.chromium.org/2667873002
Cr-Original-Commit-Position: refs/heads/master@{#448040}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b367bcc4494b1b211268529557198d77d7cc0cb0