[NTP::Cleanup] Avoid nested complex function calls
The RemoteSuggestionsProviderImpl constructor takes a long list of parameters and previously the RemoteSuggestionFetcher (which also takes a lot of parameters) was constructed in line. This was confusing, because one could not easily determine to which MakeUnique given argument belonged to.
This CL moves the RemoteSuggestionsFetcher into a local variable to improve readability.

BUG=674033

Review-Url: https://codereview.chromium.org/2671453003
Cr-Original-Commit-Position: refs/heads/master@{#448992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5a842fb2c97ce7c2e140a2d77d503c1dc6e1a299