Record Autofill form events specially for nonsecure pages
The Form-Not-Secure experiment shows special "Payment not secure" warnings for
credit card forms on HTTP pages. This CL records credit card autofill form
events in a special histogram broken out for nonsecure pages, to measure the
impact of these "Payment not secure" warnings on form interactions.

To do so, the FormEventLogger needs to know whether the main frame URL is
secure, at a point at which it doesn't have access to a form origin URL. iOS is
the only platform on which IsContextSecure used the form_origin parameter, and
there was an existing TODO to fix that (https://crbug.com/505388). So this CL
removes the form_origin parameter from IsContextSecure, fixes the iOS TODO, and
uses the resulting IsContextSecure method to record nonsecure form metrics.

BUG=677291,687823,505388

Review-Url: https://codereview.chromium.org/2672623005
Cr-Original-Commit-Position: refs/heads/master@{#449402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57464d6c82bea1c759874da3101135e1ae1494eb