Relanding "Fixed title updating for back forward navigation."
Chrome relies on KVO compliant "title" property to subscribe for
title updates. If this KVO change happens during then navigation
it is unclear if title was changes for the previous page, or the
navigation was committed and title was changed for the new page.

So if there is a navigation in progress WebController should ignore
KVO change, but it should always attempt to update title when
navigation is committed.

BUG=688047, 677356

Review-Url: https://codereview.chromium.org/2685803002
Cr-Commit-Position: refs/heads/master@{#449527}
(cherry picked from commit 487e4cfd53bac5944dad49cbce3b939a0c7bf035)

Review-Url: https://codereview.chromium.org/2691693002
Cr-Original-Commit-Position: refs/heads/master@{#449828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: adb03a4e58c33d0adc812a2ddae2180abe8ac148