Add CRWSessionEntry isEqual
EGTest have DCHECKs on CRWSessionEntry equality (e.g.
TabHistoryPopupControllerTestCase testTabHistoryMenu).
This fails because entry can be different even if they have the same
navigation item.

This is a tentative fix by the sheriff to make the tree green.
Please review this CL and revert it later if there is a better solution.

BUG=691469

Review-Url: https://codereview.chromium.org/2690973002
Cr-Original-Commit-Position: refs/heads/master@{#449923}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8d957c64f413bde7d4e32bdf375bf622446e7ff